Altitude Addend and bugfix Graduated symbology
With this PR I included a new parameter that is an added (like in the qgis2threejs plugin) to the Z value. Which is very useful to account for differences between DEM in QGIS and Google Earth DEM.
Moreover in the graduated symbology I force the colors for outside range values to be clipped.